# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MedRise Medical Centre API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /consultations:
    get:
      operationId: listConsultations
      tags: [consultations]
      summary: List consultations
      parameters:
        - name: patientId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of consultations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Consultation"
    post:
      operationId: createConsultation
      tags: [consultations]
      summary: Create a new consultation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationInput"
      responses:
        "201":
          description: Consultation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
  /consultations/{id}:
    get:
      operationId: getConsultation
      tags: [consultations]
      summary: Get a single consultation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Consultation found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
        "404":
          description: Not found
    patch:
      operationId: updateConsultation
      tags: [consultations]
      summary: Update a consultation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
        "404":
          description: Not found
    delete:
      operationId: deleteConsultation
      tags: [consultations]
      summary: Delete a consultation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── VITAL SIGNS ───────────────────────────────────────────────────────────
...
  /lab/orders:
    get:
      operationId: listLabOrders
      tags: [lab]
      summary: List lab orders
      parameters:
        - name: patientId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of lab orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LabOrder"
    post:
      operationId: createLabOrder
      tags: [lab]
      summary: Create a new lab order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabOrderInput"
      responses:
        "201":
          description: Lab order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabOrder"
  /lab/orders/{id}:
    patch:
      operationId: updateLabOrder
      tags: [lab]
      summary: Update a lab order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabOrderUpdateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabOrder"
        "404":
          description: Not found
    delete:
      operationId: deleteLabOrder
      tags: [lab]
      summary: Delete a lab order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /lab/results:
    post:
      operationId: createLabResult
      tags: [lab]
      summary: Record lab result for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabResultInput"
      responses:
        "201":
          description: Result recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabResult"
  /lab/results/{id}:
    delete:
      operationId: deleteLabResult
      tags: [lab]
      summary: Delete a lab result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── SHIFTS & LEAVE ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Consultation:
      type: object
      required: [id, patientId, visitDate, createdAt, updatedAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: ["string", "null"]
        staffId:
          type: ["integer", "null"]
        staffName:
          type: ["string", "null"]
        visitDate:
          type: string
        chiefComplaint:
          type: ["string", "null"]
        diagnosis:
          type: ["string", "null"]
        treatmentPlan:
          type: ["string", "null"]
        prescriptions:
          type: ["string", "null"]
        referral:
          type: ["string", "null"]
        followUpDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ConsultationInput:
      type: object
      required: [patientId, visitDate]
      properties:
        patientId:
          type: integer
        staffId:
          type: integer
        visitDate:
          type: string
        chiefComplaint:
          type: string
        diagnosis:
          type: string
        treatmentPlan:
          type: string
        prescriptions:
          type: string
        referral:
          type: string
        followUpDate:
          type: string
        notes:
          type: string

    # ── VITAL SIGNS SCHEMAS ───────────────────────────────────────────────
...
    LabResult:
      type: object
      required: [id, labOrderId, recordedAt]
      properties:
        id:
          type: integer
        labOrderId:
          type: integer
        result:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        referenceRange:
          type: ["string", "null"]
        interpretation:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        recordedBy:
          type: ["integer", "null"]
        recordedAt:
          type: string
    LabOrder:
      type: object
      required: [id, patientId, testName, priority, status, orderedAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: ["string", "null"]
        consultationId:
          type: ["integer", "null"]
        orderedBy:
          type: ["integer", "null"]
        orderedByName:
          type: ["string", "null"]
        testName:
          type: string
        testCategory:
          type: ["string", "null"]
        priority:
          type: string
          enum: [routine, urgent, stat]
        status:
          type: string
          enum: [pending, in-progress, completed, cancelled]
        clinicalInfo:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        orderedAt:
          type: string
        completedAt:
          type: ["string", "null"]
        results:
          type: array
          items:
            $ref: "#/components/schemas/LabResult"
    LabOrderInput:
      type: object
      required: [patientId, testName]
      properties:
        patientId:
          type: integer
        consultationId:
          type: integer
        orderedBy:
          type: integer
        testName:
          type: string
        testCategory:
          type: string
        priority:
          type: string
          enum: [routine, urgent, stat]
        clinicalInfo:
          type: string
        notes:
          type: string
    LabOrderUpdateInput:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in-progress, completed, cancelled]
        notes:
          type: string
    LabResultInput:
      type: object
      required: [labOrderId]
      properties:
        labOrderId:
          type: integer
        result:
          type: string
        unit:
          type: string
        referenceRange:
          type: string
        interpretation:
          type: string
        notes:
          type: string
        recordedBy:
          type: integer

    # ── SCHEDULE SCHEMAS ──────────────────────────────────────────────────
...