# lib/api-zod/src/generated/types/labResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MedRise Medical Centre API
 * OpenAPI spec version: 0.1.0
 */

export interface LabResult {
  id: number;
  labOrderId: number;
  /** @nullable */
  result?: string | null;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  referenceRange?: string | null;
  /** @nullable */
  interpretation?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  recordedBy?: number | null;
  recordedAt: string;
}